Blog Page Login

+++ database/migrations/2018_04_13_090907_create_sessions_table.php
@@ -15,7 +15,7 @@ class CreateSessionsTable extends Migration
     {
         Schema::create('sessions', function (Blueprint $table) {
             $table->string('id', 100)->unique();
-            $table->unsignedInteger('user_id')->nullable();
+            $table->bigInteger('user_id')->nullable();
             $table->string('ip_address', 45)->nullable();
             $table->text('user_agent')->nullable();
             $table->text('payload');
